ResourceCreator resets, Fix CVPicker buttons
CVPicker buttons were in the wrong order and clicking the text input entered
edit mode. Now, selection is on the left, creation on the righ, and input click
does not translate to edit mode.

ResourceCreator would not reset its fields when unfocused, so CV creation would
persist from element to element. This has been fixed.