Implement EditableDate, Improve decoupling
Previous work with the Modal would require flex styling for lists. No more!

createEditable now properly described as bindEditable for *binding* an editable
to a property in the store. Also removed stupid switch which used a string to
needlessly select a constructor which could be supplied as it. Also pass any
additionalProps directly into the JSX constructor.